style(llm): rename ButtonUseTouchable into ButtonUseTouchableContext
